chore(sort): fix hardcoded sort in process function
- fix the hardcoded sort in process function
- add sort options in process option
