+++ packages/sdk-client/test/client.test/sdk-client.test.ts
@@ -893,54 +893,4 @@ describe('process - exposed', () => {
       })
       .catch(fn)
   })
-
-  test('should process customer list using provided limit', async () => {
-    const projectKey = process.env.CTP_PROJECT_KEY
-    const authMiddlewareOptions = {
-      host: 'https://auth.europe-west1.gcp.commercetools.com',
-      projectKey,
-      credentials: {
-        clientId: process.env.CTP_CLIENT_ID || '',
-        clientSecret: process.env.CTP_CLIENT_SECRET || '',
-      },
-      oauthUri: process.env.adminAuthUrl || '',
-      scopes: [`manage_project:${projectKey}`],
-      fetch,
-    }
-
-    const httpMiddlewareOptions = {
-      host: 'https://api.europe-west1.gcp.commercetools.com',
-      fetch,
-    }
-
-    const client = new ClientBuilder()
-      .withProjectKey(projectKey)
-      .withClientCredentialsFlow(authMiddlewareOptions)
-      .withHttpMiddleware(httpMiddlewareOptions)
-      .build()
-
-    const apiRoot = createApiBuilderFromCtpClient(client)
-
-    const request = apiRoot
-      .withProjectKey({ projectKey })
-      .customers()
-      .get()
-      .clientRequest()
-
-    const limit = 30
-    const fn = (data: any) => data
-
-    client
-      .process(request, fn, { limit })
-      /**
-       * response is an array of processed results
-       */
-      .then((response) => {
-        // @ts-ignore
-        expect(response[0].statusCode).toEqual(200)
-        expect(response[0].body.limit).toEqual(limit)
-        expect(response[0].body.results.length).toEqual(response[0].body.count)
-      })
-      .catch(fn)
-  })
 })
